Fix nodeId generation
Using shifted nodeIds (<<) divides the id space in 2 equal sections on
every pass. This is fine if the generated tree is balanced, but if it
isn't, a large section of the id space is left reserved and unused. Thus
eventually causing overflow on the nodeId variable. Therefore we need to
pack the ids tighter.